feat: rewrite user endpoint using typeorm

+++ src/modules/user/service/user.service.ts
@@ -1,4 +1,4 @@
-import { ForbiddenException, Injectable } from '@nestjs/common';
+import { ForbiddenException, Injectable, NotFoundException } from '@nestjs/common';
 import { CreateUserDto } from '../dto/create-user.dto';
 import { UpdatePasswordDto } from '../dto/update-password.tdo';
 
@@ -14,71 +14,32 @@ export class UserService {
     private readonly usersRepository: Repository<UserEntity>,
   ) {}
 
-  public async getAllUsers(): Promise<UserEntity[]> {
-    return await this.usersRepository.find();
-    // return allUsers.map(user => user.toResponse());
+  public async getAllUsers() {
+    return (await this.usersRepository.find()).map(user => user.toResponse());
   }
 
   async createUser(userDto: CreateUserDto) {
-    const savedUser = await this.usersRepository.save({
-      // ...userDto,
-      // createdAt: +new Date(),
-      // updatedAt: +new Date(),
-    });
-    return {
-      // id: savedUser.id,
-      // login: savedUser.login,
-      // version: savedUser.version,
-      // createdAt: +savedUser.createdAt,
-      // updatedAt: +savedUser.updatedAt,
-    };
+    const createdUser = this.usersRepository.create(userDto);
+    return (await this.usersRepository.save(createdUser)).toResponse();
   }
 
   public async getUserById(id: string) {
-    const user = await this.usersRepository.findOne( { where: { id: id } } );
-    
-    if (!user) {
-      throw new ForbiddenException(`User with ${id} not found`);
-    }
-    
-    return user;
-   
-    // const currentData = {
-    //   login: user.login,
-    //   id: user.id,
-    //   createdAt: user.createdAt,
-    //   version: user.version,
-    //   updatedAt: user.updatedAt
-    // };    
-
-    // return currentData;
+    const currentUser = await this.usersRepository.findOneBy({id});
+    if (!currentUser) throw new NotFoundException(`User with ${id} not found`);
+    return currentUser.toResponse();
   }
 
-  // public async updateUser(id: string, passwords: UpdatePasswordDto): Promise<IUser> {
-  //   const index = this.inMemoryDB.users.findIndex(user => user.id === id);
-  //   const currentUser = this.inMemoryDB.users[index];
-
-  //   if (currentUser.password !== passwords.oldPassword) {
-  //     throw new ForbiddenException('Old password is incorrect');
-  //   }
-
-  //   const updatedUser = {
-  //     login: currentUser.login,
-  //     createdAt: currentUser.createdAt,
-  //     id,
-  //     version: currentUser.version + 1,
-  //     updatedAt: +Date.now()
-  //   };
-
-  //   this.inMemoryDB.users[index] = {
-  //     ...updatedUser,
-  //     password: passwords.newPassword
-  //   };
-    
-  //   return updatedUser;
-  // }
+  public async updateUser(id: string, passwords: UpdatePasswordDto) {
+    const currentUser = await this.usersRepository.findOneBy({id});
+    if (!currentUser) throw new NotFoundException(`User with ${id} not found`);
+    if (currentUser.password !== passwords.oldPassword) throw new ForbiddenException('Old password is incorrect');
+    currentUser.password = passwords.newPassword;
+    return (await this.usersRepository.save(currentUser)).toResponse();
+  }
 
   public async deleteUser(id: string): Promise<void> {
+    const currentUser = await this.usersRepository.findOneBy({id});
+    if (!currentUser) throw new NotFoundException(`User with ${id} not found`);
     await this.usersRepository.delete(id);
   }
